♻️ refactor: change api get me with add profile service
